Add files via upload
With the explosion of popular, easy-to-use messaging apps like WhatsApp, Line, Telegram, WeChat, Tango, SnapChat, Skype, Google Hangouts, Hike etc., the demand of chat apps is growing day by day. ... When it comes to building a chat app, the main advantage for businesses is: Increased Productivity.

This chat application is easy to use and used php,html,mysql and css.